pyproject.toml
[project]
name = "ai-factory"
version = "0.1.0"
description = "Self-Bootstrapping AI Company Factory"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pydantic>=2.0",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.scripts]
aifactory = "factory.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["factory*"]
